requirements.txt
# ================
# BASE
# ================
pymongo==4.15.4
numpy==2.2.6
pillow==12.0.0
tqdm==4.67.1

# ================
# TRANSFORMERS / CLIP
# ================
transformers==4.57.3
tokenizers==0.22.1
sentencepiece
regex==2025.11.3
ftfy==6.2.0
huggingface-hub==0.36.0

# ================
# MACHINE LEARNING
# ================
scikit-learn==1.7.2
scipy==1.15.3
joblib==1.5.2

# ================
# TORCH CPU (COMPATIBLE)
# ================
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1

# ================
# UTILIDADES
# ================
PyYAML==6.0.3
typing_extensions==4.15.0
